[T3b] relax sibling distanceBands schemas to positive numbers (JADE deferred to avoid a concurrent-bundle conflict)
Drops .int() from the three shared distanceBands array schemas
(pMedianInputsSchema, transportLpInputsSchema, twoEchelonInputsSchema) so a
unit conversion (e.g. 500 km -> 310.6856 mi) validates. Nothing else about
these arrays changes: they did not reject duplicate/non-ascending arrays
before this change and still do not -- the shared band classifier sorts
internally, so there is no correctness gap. Updated the corresponding band
"items" JSON-Schema entries (informational only, not re-validated at
runtime) in the four non-JADE manifests from "type": "integer" to
"type": "number", keeping "exclusiveMinimum": 0.

jadeInputs.ts and both two-echelon-jade-us/manifest.json and
chens-cosmetics-cn/manifest.json are untouched per the plan's amendment --
a concurrent bundle (workspace-fixups-2) is editing the same JADE lines
right now, so that half is deferred to a later pass to avoid a guaranteed
merge conflict.

No manifest-hash fixture needed updating: computeSha256() only hashes each
model's dataset/*.json files, never manifest.json (confirmed by reading
lib/dataset-schema/src/index.ts) -- dataset-schema tests pass unchanged.

Tests: 9 new cases (accepts non-integral bands / still rejects
zero-negative / still accepts duplicate-non-ascending) across
pMedian.test.ts, transportLp.test.ts, twoEchelon.test.ts. Full gate green:
typecheck, api-server 1015/1015, dataset-schema 38/38.

+++ artifacts/api-server/src/validation/inputs/pMedian.ts
@@ -87,7 +87,7 @@ export const pMedianInputsSchema = z.object({
   uniformCapacity: z.number().positive().nullable().optional(),
   warehouseOverrides: z.array(warehouseOverrideSchema).default([]),
   customerOverrides: z.array(customerOverrideSchema).default([]),
-  distanceBands: z.array(z.number().int().positive()).min(1),
+  distanceBands: z.array(z.number().positive()).min(1),
   gap: z.number().min(0),
   timeLimitSec: z.number().int().min(1),
   // p-median-brazil only (solve_capacitated_pmedian reads this; plain
